login by username or email feature, logout feature
